requirements.txt
redis
pytest
celery
uvicorn
fastapi
PyGithub
requests
langchain
langchain_openai
